+++ app/helpers/dorsale/expense_gun/application_helper.rb
@@ -1,6 +1,6 @@
 module Dorsale::ExpenseGun::ApplicationHelper
   def expense_states_for_filters_select
-    Dorsale::ExpenseGun::Expense.aasm.states.map do |state|
+    Dorsale::ExpenseGun::Expense::STATES.map do |state|
       [
         Dorsale::ExpenseGun::Expense.t("state.#{state}"),
         state,