+++ app/sorters/dorsale/expense_gun/expenses_sorter.rb
@@ -1,0 +1,8 @@
+class Dorsale::ExpenseGun::ExpensesSorter < Agilibox::Sorter
+  def sort
+    case column
+    when :created_at, :date
+      {column => direction}
+    end
+  end
+end