+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Next version
+- Change expenses state machine
+- Change expenses list view
 
 # 3.18.0
 - Change uploaders `extension_whitelist` to `extension_allowlist`